move checkAccess to src/lib/auth/
