+++ src/lib/core/WacLdp.ts
@@ -16,7 +16,7 @@ import { writeBlobHandler } from '../operationHandlers/writeBlobHandler'
 import { updateBlobHandler } from '../operationHandlers/updateBlobHandler'
 import { deleteBlobHandler } from '../operationHandlers/deleteBlobHandler'
 import { unknownOperationCatchAll } from '../operationHandlers/unknownOperationCatchAll'
-import { checkAccess, determineRequiredAccessModes, AccessCheckTask } from './checkAccess'
+import { checkAccess, determineRequiredAccessModes, AccessCheckTask } from '../auth/checkAccess'
 import { getAppModes } from '../auth/appIsTrustedForMode'
 import { setAppModes } from '../rdf/setAppModes'
 import { BlobTree } from '../storage/BlobTree'
